Refactor FlexArray
This avoids using an array of 1 as a flexible array member, and instead uses a
span to reference the memory after itself.
